+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 # cordova-sqlite-evplus-attach-detach-free 0.0.3-dev
 
+- feat: support ATTACH & DETACH on Windows
+
 # cordova-sqlite-evplus-attach-detach-free 0.0.2-dev
 
 ## cordova-plugin-sqlite-evplus-ext-common-free 0.8.0-rc03